% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARD.r
\docType{package}
\name{YARD}
\alias{YARD}
\alias{YARD-package}
\title{YARD: Yet Another R Demo}
\description{
This package is a learnr demo built for the West Michigan R Users Group
meeting at Big Data Ignite 2017.
}
\examples{
\dontrun{
learnr::run_tutorial("YARD", "YARD")
}
}
\author{
Paul W. Egeler, M.S., GStat
}
